requirements.txt
pydantic>=2.0.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
anthropic>=0.3.0
tenacity>=8.0.0
pyyaml>=6.0.0